html/semantics/interactive-elements/the-dialog-element/dialog-form-submission-unusual.html

<!doctype html>
<title>Test dialog form submission, unusual cases</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div id=log></div>

<dialog>
  <form method=dialog action="https://test:test/" rel=noopener>
    <input type=submit formaction="https://test:test/">
    <input type=submit>
  </form>
</dialog>

<script>
test(() => {
  const dialog = document.querySelector("dialog");
  dialog.showModal();
  assert_true(dialog.open);

  document.querySelector("input").click();
  assert_false(dialog.open);
}, "A form's action and rel=noopener are ignored during submission");

test(() => {
  const dialog = document.querySelector("dialog");
  dialog.showModal();
  assert_true(dialog.open);

  document.querySelectorAll("input")[1].click();
  assert_false(dialog.open);
}, "A form's action and rel=noopener are ignored during submission, part 2");

</script>
